backend: Use gpiozero for steppers (#572)
This completes the transition from rpi.gpio to gpiozero for the
PlanktoScope hat
Remaining code related to `m0_cs` and `m1_cs` was commented out as it
was unused and confusing.
This was tested on a PlanktoScope v2.6 and PlanktoScope v3.0

The relevant stepper code will eventually be extracted from ShushEngine
to simplify and clarify.

Partially fixes https://github.com/PlanktoScope/PlanktoScope/issues/323
(not fully as the adafruithat still uses rpi.gpio)