software: Gitify node-red-dashboard (#573)
Configure node-red to use the git repository as `userDir` so that
changes to flows are visible in git.

Here is the output of `git status` after making changed in the node-red
editor:

![Screenshot From 2025-04-16
13-36-41](https://github.com/user-attachments/assets/a9f6b68f-1a04-44ff-84ce-0aa66bdc055d)

`settings.js` file was copied from
https://github.com/PlanktoScope/pallet-standard/blob/8fd23fded73b71d391d47a14ca79684cfcb6aeb7/packages/core/apps/planktoscope/node-red-dashboard/overlays/etc/nodered/settings.js

Previously the setting file was setup using forklift
https://github.com/PlanktoScope/pallet-standard/blob/8fd23fded73b71d391d47a14ca79684cfcb6aeb7/packages/core/apps/planktoscope/node-red-dashboard/overlays/etc/systemd/system/nodered.service.d/20-override.conf

This is "bypassing" forklift as I'm not familiar with it yet, it is
unclear if it will remain and in any case [dashboard
2](https://github.com/PlanktoScope/dashboard) will likely deprecate
this.


![image](https://github.com/user-attachments/assets/8437615a-a2b4-4a15-9905-ad74c086be6a)