backend: Remove chip select related code (#579)
Before Bookworm; SPI chip select was manually managed.

With newer version of the kernel both SPI0 CE0 (GPIO 8) and SPI0 CE1
(GPIO 7) are automatically managed (and held) by the kernel.

`/dev/spidev0.0` is SPI0 CE0
`/dev/spidev0.1` is SPI0 CE1

This is why both the focus and pump are working without the chip select,
and I was able to comment it out on
https://github.com/PlanktoScope/PlanktoScope/commit/a94372a1fd38bed9b1b739871e8c19e71625a1b2

Tested on both PlanktoScope v2.6 and PlanktoScope v3 (except pump
because my prototype may be defect)

https://raspi.muth.org/spi.html