Stop building desktop images (#582)
After improving the DX, I have found that a GUI is not necessary for
development and perhaps counterproductive. The developer environment
should be as close as possible to the end product.

The idea of a desktop image was to ease connecting the Raspberry to the
network, but improvements were made in that area and I believe we should
rather focus on making sure it's easy to connect any PlanktoScope to the
network.

We now have a [developer
mode](https://github.com/PlanktoScope/PlanktoScope/blob/master/documentation/docs/community/contribute/tips-and-tricks.md#development-environment)
that can be setup on any PlanktoScope image and enable extra developer
features.

A GUI can still be useful in some scenarios, so I have documented how to
get it up and running for the [Tips and
Tricks](https://github.com/PlanktoScope/PlanktoScope/blob/master/documentation/docs/community/contribute/tips-and-tricks.md)
page.

I also believe that reducing the number of configurations (whether
officially supported or not) will allow us to iterate and improve the
PlanktoScope faster. With the end goal of building a single image.

Finally, this will (even so slightly) reduce our environmental impact.

See also this conversation
https://github.com/PlanktoScope/PlanktoScope/pull/571#issuecomment-2812235787