Add node red dashboard project (#629)
Follow up from https://github.com/PlanktoScope/PlanktoScope/pull/628

Adds [dashboard](https://github.com/PlanktoScope/dashboard/) as a
Node-RED project using a git submodule.

Why git submodule and not import `dashboard` into this repository?
Because it seems Node-RED is having issues with `git` when dealing with
higher up repos. Might fix it later but for now this will do.

Clicking on a changed file such as


![image](https://github.com/user-attachments/assets/f32875f5-4dee-4df8-b402-58c51ce62e82)

results in


![image](https://github.com/user-attachments/assets/8a975e2c-0cf3-4abe-8fdd-37a843554115)

for `planktoscopehat` and `adafruithat` projects it's fine as we're used
to using `git` cli anyway but for the new dashboard we want all the
bells and whistles of Node-RED

---

The `enable-developer-mode` script was removed as we don't clear the
`.git` folder anymore. It only saved ~100mb and clearing it is causing
headaches with submodules.