Use custom-runner for CI (#703)
Currently, we use https://github.com/ethanjli/pinspawn-action to create
and chroot into a Raspberry OS.

It's a great project but unfortunately it still limits us in several
ways

* No `firewall-cmd`
* No `systemctl enable --now` / `systemctl start`
* No `systemctl daemon-reload`
* GitHub runners doesn't provide SSH for debugging
* ...

To simplify things, let's make CI env as close as possible to the dev
env.
The dev env is the PlanktoScope itself.

Ideally CI would run on PlanktoScope so we can check for things like
pump activating and so on but for now, I've setup a Raspberry Pi 5 with
`2025-05-13-raspios-bookworm-arm64-lite.img.xz` (same base as
PlanktoScope OS) as a GitHub custom action runner.

Note that the runner is persistent, this is a feature, not a bug, this
is much faster and resource lightweight. We may clear the runners once
in a while in the future.

Eventually we'll want to build images similarly, but that'll come later.