Simplify OS setup (#662)
Remove Forklift and Docker.

## Motivation

Forklift is an interesting but experimental project. 
It takes too much kungfu to develop the PlanktoScope OS (forklift,
pallet, docker, ...)
We already have a dependency on a stable and well-supported runtime;
Raspberry Pi OS. Docker is not worth the complexity.

## Goals

The setup script should be idempotent (can run multiple times without
failing)
No unnecessary dependencies and abstractions
Better developer experience with familiar tools and quick feedback loop
Have a single env under which PlanktoScope OS is used, tested, developed
and built

## Next

Now that
1. git + just is enough to test or update a branch
2. we have a CI pipeline
https://github.com/PlanktoScope/PlanktoScope/pull/703

there is no need to build sdcard images on every commit / PR which is a
resource and time-consuming process.

However we will keep the GitHub action for releases or manual triggers.
We will stop using the systemd-nspawn method though for the reasons
explained in https://github.com/PlanktoScope/PlanktoScope/pull/703

Let's consider the qemu method or building images on real hardware.
https://jetkvm.com/ could make it easy.

Once we can run tests on a live PlanktoScope OS, we will do so. 

## Also

* eth0 is for router only, no direct connection
* Remove eth1 / wlan1 / usb interfaces
* replace device-portal with dashboard
* Replace prometheus/Grafana with PCP and cockpit-pcp