Update doc and scripts (#739)
Removes golang as `go install` is unstable and it saves space.

Update scripts to reenable/restart services. It wasn't done previously
because it breaks the image builder but we will use a different method
for building images. See
https://github.com/PlanktoScope/PlanktoScope/issues/730

We use `systemctl reenable` and `systemctl restart` instead of
`systemctl reenable --now` because of a
[bug](https://github.com/systemd/systemd/issues/31541).