Remove adafruithat support (#821)
The current state of adafruithat support is very sad in that it is
supported by being a code duplication of planktoscopehat.
All contributions are currently happening to planktoscopehat and no one
is contributing to adafruithat.

PlanktoScope OS 2025 will continue to work for adafruithat.

If anyone would like to bring back adafruithat support into `main` so
that PlanktoScope 2.1 can benefit from all the improvements that are
happening; here is what needs to be done:

1. Add support for the LED in `controller/light`
2. Add support for the Pump in `controller`
3. Add support for the focus in `controller`
4. Check what the adafruithat flow used to do that needs to be ported to
controller (fan?, ...)

FairScope / I would be happy to help if anyone is willing to work on
this. Feel free to get in touch on Slack or sonny@fairscope.com